latent_moe split, workspace lint clean, and a no-CUDA build that actually builds
Three strands, all verified against the real checkpoints before landing.

## `deepseek4/` -> `latent_moe/` + `deepseek4.rs`

The engine is now the architecture *family* and the model is one file beside
it. `latent_moe/` holds the paged/batched kernel path, provenance gallery and
wave engine; `deepseek4.rs` holds `impl Arch for DeepSeekV4` — config defaults,
GGUF metadata keys, tensor names, latent geometry, and the `dflash` drafter.
Adding a V5 is a sibling file, not a fork of the engine.

New `latent_moe/arch.rs` (the `Arch` trait plus exhaustive `Weight`/`Global`/
`Meta` enums, so a new engine tensor breaks every model at compile time) and
`latent_moe/geometry.rs` (`LatentGeometry`, const-asserted, with the kernel
tiling rules the templates depend on). Kernels are parameterised over
`<HEAD_DIM, ROPE_DIM, NPAL>` and `paged::assert_geometry` refuses a host/kernel
mismatch at load — a divergence there is wrong attention, not a fault.

`compress_ratios` had layers 40-42 as SWA when the checkpoint has CSA/HCA/CSA.
A dormant fallback that never fires is never validated; it is now asserted
element-wise against the real GGUF.

## clippy 301 -> 0, fmt clean, both feature sets

Real fixes in `src/`; blanket allows only in test/harness code, each with a
written reason. Two refusals worth recording, both NaN semantics:
`!(a > b)` -> `a <= b` and `.min().max()` -> `.clamp()` are NOT equivalent
under NaN, and the affected sites are host mirrors of CUDA kernels, so they
keep the original spelling.

Found and fixed a real bug on the way: `http_session::request` computed
`is_text_type` and then ran identical blocks in both arms, so a binary body
that happened to decode as UTF-8 came back as text.

Also four doc comments stranded on the wrong function by earlier renames and
splits (`heal_tail_divergence`, `truncate_sequence_to_blocks`, the zero-copy
rebuild block that outlived `reproject_view`, and a summariser test carrying
two `#[test]` attributes).

## The no-CUDA build

`candle-nn` and `candle-transformers` test targets had never compiled without
`cuda` — the compiler only ever checked that code in one configuration. Note
`cargo clippy --workspace` without `--features cuda` does NOT catch this:
feature unification switches `cuda` on via another member. Single-package
builds are the honest check.

The one interesting cause: `wave_plan`'s test reached `WAVE_FFN_BYTES` through
the CUDA-gated `bump_arena`. Those three constants are *measurements* — byte
budgets with no CUDA types — and everything reading them is integer
arithmetic, so they move to a new ungated `wave_spans.rs`. Same reasoning
`weight_zone` is already unconditional: provable on a machine with no GPU.

`vram_budget_available` has both a cuda and a non-cuda arm but was re-exported
only under `cuda`, so the fallback had no public path; `kv_cache/mod.rs` also
carried `#[cfg(feature = "cuda")]` twice on that line. Both fixed.

`latent_moe`, `expert_lre`, `deepseek4`, `prefill_utils`, `quantized_llama`,
`quantized_qwen2/3` and `wave_admit` are CUDA-only in substance — they name
`candle_kernels`, `cuda_backend` and the paged kernels — and are now gated to
match `quantized_qwen3_moe`, which already was. This narrows the public API
without `cuda`; safe today because `candle-conversation` pins
`features = ["cuda"]`.

## Verification

- clippy `-D warnings`: workspace+cuda = 0; per-crate no-cuda (6 crates) = 0
- `cargo fmt --all -- --check`: 0
- tests: 3,068 passed, 0 failed (cuda + cpu). The CPU column includes 427 in
  candle-nn and 58 in candle-transformers that could not run at all before.
- DeepSeek-V4-Flash batched forward, configs [1,4,8,16,1]: all valid, 100%
  session match. cfg-16 crosses 1000 t/s bulk (1032.4). Cold vs warm
  single-session 15.1 -> 20.6 t/s as expert prediction precision climbs
  46.1% -> 93.4%.
- Qwen3-30B-A3B, 16 configs: all valid. The C0-C9 ladder reproduces exactly at
  every level, 1.98x through 5.31x, with throughput flat across the range.

`coherence::codebase_tour_produces_substantial_answer` is `#[ignore]`d — it
loads the model and decodes a full tour, so it belongs in a deliberate run.

Docs relocated into `docs/archived/` and `docs/deepseek/` (content unchanged,
verified byte-for-byte against HEAD modulo line endings).

+++ candle-core/src/quantized/ko_quant.rs
@@ -347,9 +347,8 @@ pub fn quantize_ko(w: &[f32], nrows: usize, ncols: usize, dtype: GgmlDType) -> V
                         if crumb_bytes > 0 {
                             let (mut cr0, mut cr1) = (0u8, 0u8);
                             for j in 0..4 {
-                                cr0 |= (((q[sub * 32 + q3 * 4 + j] >> 4) & 0x3) << (2 * j)) as u8;
-                                cr1 |=
-                                    (((q[sub * 32 + 16 + q3 * 4 + j] >> 4) & 0x3) << (2 * j)) as u8;
+                                cr0 |= ((q[sub * 32 + q3 * 4 + j] >> 4) & 0x3) << (2 * j);
+                                cr1 |= ((q[sub * 32 + 16 + q3 * 4 + j] >> 4) & 0x3) << (2 * j);
                             }
                             ob[cbase + crumb_base + lane * 8 + sub * 2] = cr0;
                             ob[cbase + crumb_base + lane * 8 + sub * 2 + 1] = cr1;
@@ -357,8 +356,8 @@ pub fn quantize_ko(w: &[f32], nrows: usize, ncols: usize, dtype: GgmlDType) -> V
                         if hi_bytes > 0 {
                             let (mut hb0, mut hb1) = (0u8, 0u8);
                             for j in 0..4 {
-                                hb0 |= (((q[sub * 32 + q3 * 4 + j] >> 4) & 1) << j) as u8;
-                                hb1 |= (((q[sub * 32 + 16 + q3 * 4 + j] >> 4) & 1) << j) as u8;
+                                hb0 |= ((q[sub * 32 + q3 * 4 + j] >> 4) & 1) << j;
+                                hb1 |= ((q[sub * 32 + 16 + q3 * 4 + j] >> 4) & 1) << j;
                             }
                             ob[cbase + hi_base + lane * 4 + sub] = hb0 | (hb1 << 4);
                         }
